Add files via upload
first draft of the chapter 2 analysis– it's based on the chapter 3 analysis, so it includes several chunks of copied code. The sections I have put my hands on for chapter 2 are "Data", which now correctly excludes participants down to the correct ids (n = 84). the pca code chunk is also edited to suit the pca loadings for chapter 2. the distribution of boundary response graph is also edited to suit chapter 2. The following code chunks are still a WIP.